Add cappigen to the install targets.
git-svn-id: http://svn.eol.ucar.edu/svn/eldora/trunk@2636 15f24ed5-0439-0410-88a4-bd1d0fc06cdd

+++ src/cappi_gen/CappiGen.cpp
@@ -153,7 +153,7 @@ void CappiGen::productSlot(
     }
 
     // copy the product into an array, and write to output file
-    double product[_maxCells];
+    float product[_maxCells];
     //    std::cerr << "copying " << numCells << " gates" << std::endl;
     int i;
     for (i = 0; i < numCells; ++i) {
